V11/requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.25
python-dotenv==1.0.0
PyQt6==6.6.1
qasync==0.27.1
pandas==2.2.0
openpyxl==3.1.2
qtawesome==1.3.1
matplotlib==3.8.2
arabic-reshaper==3.0.0
python-bidi==0.4.2
aiohttp==3.9.1
colorlog==6.8.0
Pillow==10.2.0